+++ common/src/main/kotlin/dev/vanutp/tgbridge/common/DefaultConfigUnchangedException.kt
@@ -1,0 +1,3 @@
+package dev.vanutp.tgbridge.common
+
+class DefaultConfigUnchangedException : Exception("botToken or chatId is not set")
