+++ common/src/main/kotlin/dev/vanutp/tgbridge/common/TelegramBot.kt
@@ -1,210 +1,19 @@
 package dev.vanutp.tgbridge.common
 
-import com.google.gson.annotations.SerializedName
+import dev.vanutp.tgbridge.common.dataclass.*
+import dev.vanutp.tgbridge.common.models.TgApi
 import kotlinx.coroutines.*
 import okhttp3.OkHttpClient
 import retrofit2.HttpException
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.http.Body
-import retrofit2.http.GET
-import retrofit2.http.POST
-import retrofit2.http.Query
 import java.time.Duration
 
-data class TgUser(
-    val id: Long,
-    @SerializedName("first_name")
-    val firstName: String,
-    @SerializedName("last_name")
-    val lastName: String?,
-    val username: String?,
-)
-
-data class TgChat(
-    val id: Long,
-    val title: String = "",
-    val username: String? = null,
-)
-
-class TgAny
-
-data class TgPoll(
-    val question: String,
-)
-
-data class TgMessageOrigin(
-    @SerializedName("sender_user")
-    val senderUser: TgUser? = null,
-    @SerializedName("sender_user_name")
-    val senderUserName: String? = null,
-    @SerializedName("sender_chat")
-    val senderChat: TgChat? = null,
-    val chat: TgChat? = null,
-)
-
-interface TgMessageMedia {
-    val animation: TgAny?
-    val photo: List<TgAny>?
-    val audio: TgAny?
-    val document: TgAny?
-    val sticker: TgAny?
-    val video: TgAny?
-    val videoNote: TgAny?
-    val voice: TgAny?
-    val poll: TgPoll?
-}
-
-data class TgExternalReplyInfo(
-    val origin: TgMessageOrigin,
-    val chat: TgChat? = null,
-    override val animation: TgAny? = null,
-    override val photo: List<TgAny>? = null,
-    override val audio: TgAny? = null,
-    override val document: TgAny? = null,
-    override val sticker: TgAny? = null,
-    override val video: TgAny? = null,
-    @SerializedName("video_note")
-    override val videoNote: TgAny? = null,
-    override val voice: TgAny? = null,
-    override val poll: TgPoll? = null,
-) : TgMessageMedia {
-    val senderName
-        get() = origin.senderUser?.let {
-            (it.firstName + " " + (it.lastName ?: "")).trim()
-        }
-            ?: origin.senderUserName
-            ?: origin.senderChat?.title
-            ?: origin.chat?.title
-            ?: ""
-}
-
-data class TgTextQuote(
-    val text: String,
-)
-
-data class TgMessage(
-    val chat: TgChat,
-    @SerializedName("message_id")
-    val messageId: Int,
-    val from: TgUser? = null,
-    @SerializedName("sender_chat")
-    val senderChat: TgChat? = null,
-    @SerializedName("forward_from")
-    val forwardFrom: TgUser? = null,
-    @SerializedName("forward_from_chat")
-    val forwardFromChat: TgChat? = null,
-    @SerializedName("reply_to_message")
-    val replyToMessage: TgMessage? = null,
-    @SerializedName("external_reply")
-    val externalReply: TgExternalReplyInfo? = null,
-    val quote: TgTextQuote? = null,
-    @SerializedName("message_thread_id")
-    val messageThreadId: Int? = null,
-    @SerializedName("author_signature")
-    val authorSignature: String? = null,
-    val text: String? = null,
-    val caption: String? = null,
-    override val animation: TgAny? = null,
-    override val photo: List<TgAny>? = null,
-    override val audio: TgAny? = null,
-    override val document: TgAny? = null,
-    override val sticker: TgAny? = null,
-    override val video: TgAny? = null,
-    @SerializedName("video_note")
-    override val videoNote: TgAny? = null,
-    override val voice: TgAny? = null,
-    override val poll: TgPoll? = null,
-    @SerializedName("pinned_message")
-    val pinnedMessage: TgMessage? = null,
-) : TgMessageMedia {
-    val senderName
-        get() = authorSignature
-            ?: senderChat?.title
-            ?: from?.let {
-                (it.firstName + " " + (it.lastName ?: "")).trim()
-            }
-            ?: ""
-    val senderUserName
-        get() = from?.username
-            ?: senderChat?.username
-            ?: ""
-    val effectiveText
-        get() = text ?: caption
-}
-
-data class TgUpdate(
-    @SerializedName("update_id")
-    val updateId: Int,
-    val message: TgMessage? = null,
-)
-
-data class TgResponse<T>(
-    val ok: Boolean,
-    val result: T?,
-    val description: String?,
-)
-
-data class TgSendMessageRequest(
-    @SerializedName("chat_id")
-    val chatId: Long,
-    @SerializedName("text")
-    val text: String,
-    @SerializedName("reply_to_message_id")
-    val replyToMessageId: Int? = null,
-    @SerializedName("parse_mode")
-    val parseMode: String = "HTML",
-    @SerializedName("disable_web_page_preview")
-    val disableWebPagePreview: Boolean = true,
-)
-
-data class TgEditMessageRequest(
-    @SerializedName("chat_id")
-    val chatId: Long,
-    @SerializedName("message_id")
-    val messageId: Int,
-    @SerializedName("text")
-    val text: String,
-    @SerializedName("parse_mode")
-    val parseMode: String = "HTML",
-    @SerializedName("disable_web_page_preview")
-    val disableWebPagePreview: Boolean = true,
-)
-
-data class TgDeleteMessageRequest(
-    @SerializedName("chat_id")
-    val chatId: Long,
-    @SerializedName("message_id")
-    val messageId: Int,
-)
-
-interface TgApi {
-    @GET("getMe")
-    suspend fun getMe(): TgResponse<TgUser>
-
-    @POST("sendMessage")
-    suspend fun sendMessage(@Body data: TgSendMessageRequest): TgResponse<TgMessage>
-
-    @POST("editMessageText")
-    suspend fun editMessageText(@Body data: TgEditMessageRequest): TgResponse<TgMessage>
-
-    @POST("deleteMessage")
-    suspend fun deleteMessage(@Body data: TgDeleteMessageRequest): TgResponse<Boolean>
-
-    @GET("getUpdates")
-    suspend fun getUpdates(
-        @Query("offset") offset: Int,
-        @Query("timeout") timeout: Int,
-        @Query("allowed_updates") allowedUpdates: List<String> = listOf("message"),
-    ): TgResponse<List<TgUpdate>>
-
-    @POST("deleteWebhook")
-    suspend fun deleteWebhook(): TgResponse<Boolean>
-}
-
-const val POLL_TIMEOUT_SECONDS = 60
 
 class TelegramBot(private val botToken: String, private val logger: AbstractLogger) {
+
+    private val POLL_TIMEOUT_SECONDS = 60
+
     private val okhttpClient = OkHttpClient.Builder()
         .readTimeout(Duration.ofSeconds((POLL_TIMEOUT_SECONDS + 10).toLong()))
         .build()
