+++ common/src/main/kotlin/dev/vanutp/tgbridge/common/dataclass/TgAny.kt
@@ -1,0 +1,3 @@
+package dev.vanutp.tgbridge.common.dataclass
+
+class TgAny
