+++ common/src/main/kotlin/dev/vanutp/tgbridge/common/dataclass/TgChat.kt
@@ -1,0 +1,7 @@
+package dev.vanutp.tgbridge.common.dataclass
+
+data class TgChat(
+    val id: Long,
+    val title: String = "",
+    val username: String? = null,
+)
