+++ common/src/main/kotlin/dev/vanutp/tgbridge/common/dataclass/TgDeleteMessageRequest.kt
@@ -1,0 +1,10 @@
+package dev.vanutp.tgbridge.common.dataclass
+
+import com.google.gson.annotations.SerializedName
+
+data class TgDeleteMessageRequest(
+    @SerializedName("chat_id")
+    val chatId: Long,
+    @SerializedName("message_id")
+    val messageId: Int,
+)
