+++ common/src/main/kotlin/dev/vanutp/tgbridge/common/dataclass/TgPoll.kt
@@ -1,0 +1,5 @@
+package dev.vanutp.tgbridge.common.dataclass
+
+data class TgPoll(
+    val question: String,
+)
