+++ common/src/main/kotlin/dev/vanutp/tgbridge/common/dataclass/TgResponse.kt
@@ -1,0 +1,7 @@
+package dev.vanutp.tgbridge.common.dataclass
+
+data class TgResponse<T>(
+    val ok: Boolean,
+    val result: T?,
+    val description: String?,
+)
