+++ common/src/main/kotlin/dev/vanutp/tgbridge/common/dataclass/TgTextQuote.kt
@@ -1,0 +1,5 @@
+package dev.vanutp.tgbridge.common.dataclass
+
+data class TgTextQuote(
+    val text: String,
+)
