+++ common/src/main/kotlin/dev/vanutp/tgbridge/common/dataclass/TgUpdate.kt
@@ -1,0 +1,9 @@
+package dev.vanutp.tgbridge.common.dataclass
+
+import com.google.gson.annotations.SerializedName
+
+data class TgUpdate(
+    @SerializedName("update_id")
+    val updateId: Int,
+    val message: TgMessage? = null,
+)
