+++ common/src/main/kotlin/dev/vanutp/tgbridge/common/dataclass/TgUser.kt
@@ -1,0 +1,12 @@
+package dev.vanutp.tgbridge.common.dataclass
+
+import com.google.gson.annotations.SerializedName
+
+data class TgUser(
+    val id: Long,
+    @SerializedName("first_name")
+    val firstName: String,
+    @SerializedName("last_name")
+    val lastName: String?,
+    val username: String?,
+)
